rag-backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Document Processing (use latest for pre-built wheels)
pymupdf>=1.24.0
python-docx>=1.1.0
markdown>=3.5.2
chardet>=5.2.0

# Text Processing & Chunking
tiktoken>=0.5.2

# Embeddings
sentence-transformers>=2.3.1

# Vector Database (ChromaDB - local, easy setup)
chromadb>=0.4.22

# LLM Providers
google-generativeai>=0.4.0
openai>=1.12.0

# Utilities
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
aiofiles>=23.2.1

# Authentication
python-jose[cryptography]>=3.3.0

# Rate Limiting
slowapi>=0.1.9

# Monitoring & Metrics
prometheus-client>=0.19.0
prometheus-fastapi-instrumentator>=6.1.0

# Database
sqlalchemy>=2.0.23

# Evaluation & Testing
pytest>=8.0.0
httpx>=0.26.0

